+++ tests/tests.py
@@ -1,6 +1,9 @@
-from sessions import Session
-from games import Game
-from players import Player
+import os, sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+from baddy.sessions import Session
+from baddy.games import Game
+from baddy.players import Player
 
 # sessions
 test_session = Session(weekday='Tuesday', session='6:30')
@@ -45,6 +48,5 @@ test_session.players[0].name
 test_session.players[0].tier
 
 # Games can be created
-
 test_session.add_game(test_game)
 test_session.games[0]
